+++ st2actions/st2actions/runners/localrunner.py
@@ -80,8 +80,6 @@ class LocalShellRunner(ActionRunner, ShellRunnerMixin):
                                                    LOCAL_RUNNER_DEFAULT_ACTION_TIMEOUT)
 
     def run(self, action_parameters):
-        LOG.debug('    action_parameters = %s', action_parameters)
-
         env_vars = self._env
 
         if not self.entry_point:
@@ -169,5 +167,4 @@ class LocalShellRunner(ActionRunner, ShellRunnerMixin):
             result['error'] = error
 
         status = LIVEACTION_STATUS_SUCCEEDED if exit_code == 0 else LIVEACTION_STATUS_FAILED
-        self._log_action_completion(logger=LOG, result=result, status=status, exit_code=exit_code)
         return (status, jsonify.json_loads(result, LocalShellRunner.KEYS_TO_TRANSFORM), None)