+++ st2actions/tests/unit/test_queue_consumers.py
@@ -19,7 +19,8 @@ tests_config.parse_args()
 import mock
 from kombu.message import Message
 
-from st2actions import scheduler, worker
+from st2actions import worker
+from st2actions import scheduler
 from st2actions.container.base import RunnerContainer
 from st2common.constants import action as action_constants
 from st2common.models.db.liveaction import LiveActionDB