+++ st2reactor/st2reactor/sensor/config.py
@@ -21,6 +21,19 @@ from st2common.constants.system import VERSION_STRING
 CONF = cfg.CONF
 
 
+def parse_args(args=None):
+    CONF(args=args, version=VERSION_STRING)
+
+
+def register_opts(ignore_errors=False):
+    _register_common_opts(ignore_errors=ignore_errors)
+    _register_sensor_container_opts(ignore_errors=ignore_errors)
+
+
+def get_logging_config_path():
+    return cfg.CONF.sensorcontainer.logging
+
+
 def _register_common_opts(ignore_errors=False):
     st2cfg.register_opts(ignore_errors=ignore_errors)
 
@@ -32,7 +45,8 @@ def _register_sensor_container_opts(ignore_errors=False):
     ]
     st2cfg.do_register_opts(logging_opts, group='sensorcontainer', ignore_errors=ignore_errors)
 
-    sensor_test_opt = cfg.StrOpt('sensor-name', help='Only run sensor with the provided name.')
+    sensor_test_opt = cfg.StrOpt('sensor-ref', help='Only run sensor with the provided reference. \
+        Value is of the form pack.sensor-name.')
     st2cfg.do_register_cli_opts(sensor_test_opt, ignore_errors=ignore_errors)
 
     st2_webhook_opts = [
@@ -52,13 +66,4 @@ def _register_sensor_container_opts(ignore_errors=False):
                             ignore_errors=ignore_errors)
 
 
-def register_opts(ignore_errors=False):
-    _register_common_opts(ignore_errors=ignore_errors)
-    _register_sensor_container_opts(ignore_errors=ignore_errors)
-
-
 register_opts(ignore_errors=True)
-
-
-def parse_args(args=None):
-    CONF(args=args, version=VERSION_STRING)