+++ CHANGELOG.rst
@@ -18,6 +18,14 @@ in development
   bug-fix)
 * Allow users to use ``timediff_lt`` and ``timediff_gt`` rule comparison operator with many string
   date formats - previously it only worked with ISO8601 date strings. (improvement)
+* Allow user to specify new ``secret`` attribute (boolean) for each action parameters. Values of
+  parameters which have this attribute set to true will be masked in the log files. (new-feature)
+* API server now gracefully shuts down on SIGINT (CTRL-C). (improvement)
+* Fix a bug with with reinstalling a pack with no existing config - only try to move the config
+  file over if it exists. (bug fix)
+* Support for masking secret parameters in the API responses. Secret parameters can only be viewed
+  through the API by admin users. (new-feature)
+* Single sensor mode of Sensor Container uses ``--sensor-ref`` instead of ``--sensor-name``.
 
 0.11.3 - June 16, 2015
 ----------------------